requirements.txt
# RoboCupRescue RMRC 2026 - Python Requirements
# ==============================================

# Laptop (Operator Station)
# -------------------------
PyQt5>=5.15.0
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
pygame>=2.1.0
pyzmq>=25.0.0
msgpack>=1.0.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
reportlab>=3.6.0

# Jetson (Robot Brain) - also install these on Jetson
# ---------------------------------------------------
# pyzmq>=25.0.0
# numpy>=1.21.0
# opencv-python-headless>=4.5.0  # Use headless on Jetson
# pyserial>=3.5
# pyzbar>=0.1.9
# ultralytics>=8.0.0  # YOLOv8
# scipy>=1.7.0
# transforms3d>=0.4.1

# Development
# -----------
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
